# packages/mail-api/openapi/mail-proxy.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mail Proxy API",
    "version": "0.1.0",
    "description": "REST facade over Mailcow IMAP/SMTP and SOGo CalDAV"
  },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "CreateCalendarRequest": {
        "type": "object",
        "required": ["displayName"],
        "properties": {
          "displayName": { "type": "string" },
          "color": { "type": "string", "nullable": true }
        }
      },
      "UpdateCalendarRequest": {
        "type": "object",
        "properties": {
          "displayName": { "type": "string" },
          "color": { "type": "string", "nullable": true }
        }
      },
      "MoveCalendarEventRequest": {
        "type": "object",
        "required": ["fromCalendarId", "toCalendarId"],
        "properties": {
          "fromCalendarId": { "type": "string" },
          "toCalendarId": { "type": "string" },
          "recurrenceId": { "type": "string", "nullable": true }
        }
      },
      "CalendarEventScope": {
        "type": "string",
        "enum": ["this", "thisAndFuture", "all"]
      },
      "FreeBusySlot": {
        "type": "object",
        "required": ["start", "end"],
        "properties": {
          "start": { "type": "string", "format": "date-time" },
          "end": { "type": "string", "format": "date-time" }
        }
      },
      "FreeBusyEntry": {
        "type": "object",
        "required": ["email", "busy"],
        "properties": {
          "email": { "type": "string" },
          "busy": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FreeBusySlot" }
          }
        }
      },
      "FreeBusyResponse": {
        "type": "object",
        "required": ["entries"],
        "properties": {
          "entries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FreeBusyEntry" }
          }
        }
      },
      "CalendarImportRequest": {
        "type": "object",
        "required": ["calendarId", "ics"],
        "properties": {
          "calendarId": { "type": "string" },
          "ics": { "type": "string" }
        }
      },
      "CalendarInfo": {
        "type": "object",
        "required": ["id", "displayName", "color", "ctag"],
        "properties": {
          "id": { "type": "string" },
          "displayName": { "type": "string" },
          "color": { "type": "string", "nullable": true },
          "ctag": { "type": "string", "nullable": true }
        }
      },
...
      "CalendarIcsResource": {
        "type": "object",
        "required": ["calendarId", "ics"],
        "properties": {
          "calendarId": { "type": "string" },
          "etag": { "type": "string", "nullable": true },
          "ics": { "type": "string" }
        }
      },
...
      "CalendarIcsRequest": {
        "type": "object",
        "required": ["calendarId", "ics"],
        "properties": {
          "calendarId": { "type": "string" },
          "ics": { "type": "string" },
          "etag": { "type": "string" },
          "recurrenceId": { "type": "string", "nullable": true },
          "scope": { "$ref": "#/components/schemas/CalendarEventScope" }
        }
      },
      "CalendarsResponse": {
        "type": "object",
        "required": ["calendars"],
        "properties": {
          "calendars": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CalendarInfo" }
          }
        }
      },
      "CalendarEventsResponse": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CalendarIcsResource" }
          }
        }
      },
      "CalendarEventResponse": {
        "$ref": "#/components/schemas/CalendarIcsResource"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/calendar/calendars": {
      "get": {
        "operationId": "listCalendars",
        "summary": "List calendars",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Calendar list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "createCalendar",
        "summary": "Create calendar",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCalendarRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Calendar created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/calendars/{calendarId}": {
      "patch": {
        "operationId": "updateCalendar",
        "summary": "Update calendar",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCalendarRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calendar updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "operationId": "deleteCalendar",
        "summary": "Delete calendar",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Calendar deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/events/search": {
      "get": {
        "operationId": "searchCalendarEvents",
        "summary": "Search calendar events",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          {
            "name": "calendarId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/events/import": {
      "post": {
        "operationId": "importCalendarEvent",
        "summary": "Import calendar event from ICS",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CalendarImportRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event imported",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/freebusy": {
      "get": {
        "operationId": "queryCalendarFreeBusy",
        "summary": "Query free/busy",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "emails",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Free/busy data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FreeBusyResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/events/{eventUid}/move": {
      "post": {
        "operationId": "moveCalendarEvent",
        "summary": "Move event to another calendar",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventUid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MoveCalendarEventRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event moved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/events/{eventUid}/export": {
      "get": {
        "operationId": "exportCalendarEvent",
        "summary": "Export event as ICS",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventUid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "calendarId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "ICS file",
            "content": {
              "text/calendar": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/calendar/events": {
      "get": {
        "operationId": "queryCalendarEvents",
        "summary": "Query calendar events",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "calendarId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Comma-separated calendar IDs"
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Events in range",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "createCalendarEvent",
        "summary": "Create calendar event",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CalendarIcsRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/calendar/events/{eventUid}": {
      "get": {
        "operationId": "getCalendarEvent",
        "summary": "Get calendar event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventUid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "calendarId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "recurrenceId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Event detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "updateCalendarEvent",
        "summary": "Update calendar event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventUid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CalendarIcsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CalendarEventResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "operationId": "deleteCalendarEvent",
        "summary": "Delete calendar event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventUid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "calendarId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "recurrenceId", "in": "query", "schema": { "type": "string" } },
          {
            "name": "scope",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/CalendarEventScope" }
          }
        ],
        "responses": {
          "204": { "description": "Event deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  }
}